# nestjs-project/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/me/subscriptions": {
      "get": {
        "description": "Paginated listing of the caller's followed channels (per social-interactions/TD-05).",
        "operationId": "UsersController_getMySubscriptions",
        "parameters": [
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Pagination offset (default 0)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Max channels to return (default 20, max 50)",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated followed channels",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "nickname": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "avatarUrl": {
                            "type": "string",
                            "nullable": true
                          }
                        }
                      }
                    },
                    "total": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List channels the caller follows",
        "tags": [
          "users"
        ]
      }
    },
    "/channels/me": {
      "get": {
        "description": "Returns the caller's own channel.",
        "operationId": "ChannelsController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Own channel",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "name": {
                      "type": "string"
                    },
                    "nickname": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string",
                      "nullable": true
                    },
                    "created_at": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updated_at": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get own channel",
        "tags": [
          "channels"
        ]
      },
      "patch": {
        "description": "Updates nickname, name, and/or description of the caller's own channel, rejecting with 409 when the requested nickname is already in use (per phase-04-video-channel-management/TD-04).",
        "operationId": "ChannelsController_updateMe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateChannelDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated channel",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "name": {
                      "type": "string"
                    },
                    "nickname": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string",
                      "nullable": true
                    },
                    "created_at": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updated_at": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Requested nickname is already in use",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update own channel",
        "tags": [
          "channels"
        ]
      }
    },
...
    "/channels/{nickname}/videos": {
      "get": {
        "description": "Paginated listing of a channel's published, public videos (per phase-04-video-channel-management/TD-02, TD-05).",
        "operationId": "ChannelsController_getChannelVideos",
        "parameters": [
          {
            "name": "nickname",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "latest",
                "popular",
                "oldest"
              ],
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated video list",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "public_id": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string",
                            "nullable": true
                          },
                          "thumbnail_key": {
                            "type": "string",
                            "nullable": true
                          },
                          "duration_seconds": {
                            "type": "number",
                            "nullable": true
                          },
                          "published_at": {
                            "type": "string",
                            "format": "date-time",
                            "nullable": true
                          },
                          "views": {
                            "type": "number"
                          }
                        }
                      }
                    },
                    "page": {
                      "type": "number"
                    },
                    "limit": {
                      "type": "number"
                    },
                    "total": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No channel with this nickname",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          }
        },
        "summary": "List a channel's public videos",
        "tags": [
          "channels"
        ]
      }
    },
...
  },
  "info": {
    "title": "StreamTube API",
    "description": "API REST do StreamTube",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateChannelDto": {
        "type": "object",
        "properties": {}
      },
...
      "ApiErrorEnvelope": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 401
          },
          "error": {
            "type": "string",
            "example": "INVALID_CREDENTIALS"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": "Invalid email or password"
          },
          "code": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "statusCode",
          "error",
          "message"
        ]
      }
    }
  }
}